+++ utils/catalog-parser/src/main/java/de/sovity/edc/utils/catalog/mapper/DspDataOfferBuilder.java
@@ -13,6 +13,7 @@
 
 package de.sovity.edc.utils.catalog.mapper;
 
+import de.sovity.edc.utils.JsonUtils;
 import de.sovity.edc.utils.catalog.DspCatalogServiceException;
 import de.sovity.edc.utils.catalog.model.DspCatalog;
 import de.sovity.edc.utils.catalog.model.DspContractOffer;
@@ -21,14 +22,14 @@ import de.sovity.edc.utils.jsonld.JsonLdUtils;
 import de.sovity.edc.utils.jsonld.vocab.Prop;
 import jakarta.json.Json;
 import jakarta.json.JsonObject;
-import jakarta.json.JsonString;
 import lombok.RequiredArgsConstructor;
 import lombok.val;
+import org.eclipse.edc.connector.contract.spi.ContractId;
 import org.eclipse.edc.jsonld.spi.JsonLd;
 import org.eclipse.edc.spi.monitor.Monitor;
 import org.jetbrains.annotations.NotNull;
 
-import java.io.StringWriter;
+import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.util.Base64;
@@ -77,47 +78,62 @@ public class DspDataOfferBuilder {
     DspContractOffer buildContractOffer(JsonObject json) {
         /*
          * /!\ Workaround
-         * TODO: can't reference a private repo in a public repo
-         * https://github.com/sovity/edc-broker-server-extension/issues/278
-         * https://github.com/sovity/edc-broker-server-extension/issues/409
          *
-         * The Eclipse EDC uses a new random ID for each contract offer that it returns.
-         * This can't be used as an id.
+         * The Eclipse EDC uses a new random UUID for each policy that it returns and in turn a new contract ID.
+         * This Eclipse ID can't be used as such.
          * As a workaround, we must introduce our own ID.
          * For a first iteration, we will assume that the content of the policy remains the same (same content, same order)
          * and hash it to use it as a key.
          */
 
-        String idFieldName = "@id";
-        val id = json.get(idFieldName);
-        val idAsString = (JsonString) id;
-        val parts = idAsString.getString().split(":");
-        if (parts.length != 3) {
-            throw new RuntimeException("Can't use " + idAsString + ": wrong format, must be made of 3 parts.");
+        // FIXME: This doesn't enforce any property order and may cause trouble if the returned policy schema is not consistent
+        //  Use canonical form if needed later.
+        val noId = Json.createObjectBuilder(json).remove(Prop.ID).build();
+        val policyJsonString = JsonUtils.toJson(noId);
+        val sha1 = sha1(policyJsonString);
+        // encoding with base16 to make the hash readable to humans when decoding with ContractId
+        val humanReadableSha1 = toBase16String(sha1);
+        // re-encoding as base64 because this is the format that ContractId expects
+        val policyId = toBase64String(humanReadableSha1);
+
+        val idAsString = JsonLdUtils.string(json, Prop.ID);
+
+        val stableId = ContractId
+                .parseId(idAsString)
+                .map(it -> it.definitionPart() + ":" + it.assetIdPart() + ":" + policyId)
+                .orElseThrow((failure) -> {
+                    String message = "Failed to parse the contract id: " + failure.getFailureDetail();
+                    monitor.severe(message);
+                    throw new RuntimeException(message);
+                });
+
+        val copy = Json.createObjectBuilder(json).remove(Prop.ID).add(Prop.ID, stableId).build();
+
+        return new DspContractOffer(stableId, copy);
+    }
+
+    private static String toBase64String(String string) {
+        byte[] stringBytes = string.getBytes(StandardCharsets.UTF_8);
+        byte[] bytes = Base64.getEncoder().encode(stringBytes);
+        return new String(bytes);
+    }
+
+    @NotNull
+    private static String toBase16String(byte[] bytes) {
+        val sb = new StringBuilder();
+        for (byte b : bytes) {
+            sb.append(Character.forDigit(b >> 4 & 0xf, 16));
+            sb.append(Character.forDigit(b & 0xf, 16));
         }
+        return sb.toString();
+    }
 
-        val sw = new StringWriter();
-        try (val writer = Json.createWriter(sw)) {
-            // FIXME: This doesn't enforce any property order and may cause trouble if the returned policy schema is not consistent
-            //  Use canonical form if needed later.
-            val noId = Json.createObjectBuilder(json).remove(idFieldName).build();
-            writer.write(noId);
-            val policyJsonString = sw.toString();
-            try {
-                val hash = MessageDigest.getInstance("sha-1").digest(policyJsonString.getBytes());
-                val b64 = Base64.getEncoder().encode(hash);
-                val contractId = parts[0];
-                val assetId = parts[1];
-                val policyId = new String(b64);
-                val stableId = contractId + ":" + assetId + ":" + policyId;
-
-                val copy = Json.createObjectBuilder(json).remove(idFieldName).add(idFieldName, stableId).build();
-
-                return new DspContractOffer(stableId, copy);
-            } catch (NoSuchAlgorithmException e) {
-                monitor.severe("Failed to hash with sha-1", e);
-                throw new RuntimeException(e);
-            }
+    private byte[] sha1(String policyJsonString) {
+        try {
+            return MessageDigest.getInstance("sha-1").digest(policyJsonString.getBytes());
+        } catch (NoSuchAlgorithmException e) {
+            monitor.severe("Failed to hash with sha-1", e);
+            throw new RuntimeException(e);
         }
     }
 }